src/requirements.txt
sty
rich
pygal
numpy
pandas
bcrypt
seaborn
xgboost
lightgbm
markdown
matplotlib
scikit-learn
python-dotenv
scikit-optimize
google-generativeai
Flask
Flask-Dance
Flask-RESTful
Flask-Session
Flask-SocketIO
Flask-SQLAlchemy
